V1
V1. just look at the site to know what changed. this is just all of the basic functionality being implement with a good enough UI/UX. will come back some day to this, or not (prob not cuz wouldn't be at Brébeuf for ever but still, was fun).
Could very much be the last commit before this gets archived in about a year or whenever I decide to sort through my repos.
さよなら！